// src/api/generated/schema.d.ts
export interface components {
    schemas: {
        BMI: {
            readonly id?: number;
            child: number;
            bmi: number;
            /** Format: date */
            date?: string;
            notes?: string | null;
            tags?: string[];
        };
        DiaperChange: {
            readonly id?: number;
            child: number;
            /** Format: date-time */
            time?: string;
            wet: boolean;
            solid: boolean;
            /** @enum {string} */
            color?: "black" | "brown" | "green" | "yellow";
            amount?: number | null;
            notes?: string | null;
            tags?: string[];
        };
        Child: {
            readonly id?: number;
            first_name: string;
            last_name?: string;
            /** Format: date */
            birth_date: string;
            birth_time?: string | null;
            readonly slug?: string;
            /** Format: binary */
            picture?: string | null;
        };
        Feeding: {
            readonly id?: number;
            /** @description Required unless a Timer value is provided. */
            child?: number | null;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            start?: string;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            end?: string;
            /** @description May be used in place of the Start, End, and/or Child values. */
            timer?: number | null;
            readonly duration?: string | null;
            /** @enum {string} */
            type: "breast milk" | "formula" | "fortified breast milk" | "solid food";
            /** @enum {string} */
            method: "bottle" | "left breast" | "right breast" | "both breasts" | "parent fed" | "self fed";
            amount?: number | null;
            notes?: string | null;
            tags?: string[];
        };
// ...
        Medication: {
            readonly id?: number;
            child: number;
            /** @description Name of the medication administered */
            name: string;
            /** @description Amount of medication given */
            dosage?: number | null;
            /** @enum {string} */
            dosage_unit?: "mg" | "ml" | "tablets" | "drops";
            /** Format: date-time */
            time?: string;
            /** @description Time until next dose can be given */
            next_dose_interval?: string | null;
            notes?: string | null;
            tags?: string[];
        };
// ...
        Sleep: {
            readonly id?: number;
            /** @description Required unless a Timer value is provided. */
            child?: number | null;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            start?: string;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            end?: string;
            /** @description May be used in place of the Start, End, and/or Child values. */
            timer?: number | null;
            readonly duration?: string | null;
            nap?: boolean | null;
            notes?: string | null;
            tags?: string[];
        };
// ...
        TummyTime: {
            readonly id?: number;
            /** @description Required unless a Timer value is provided. */
            child?: number | null;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            start?: string;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            end?: string;
            /** @description May be used in place of the Start, End, and/or Child values. */
            timer?: number | null;
            readonly duration?: string | null;
            milestone?: string;
            tags?: string[];
        };
        Weight: {
            readonly id?: number;
            child: number;
            weight: number;
            /** Format: date */
            date?: string;
            notes?: string | null;
            tags?: string[];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}